move load_yaml function and related class to new utilities module
